Put RitualStates in a common spot
